\docType{package}
\name{oct4-package}
\alias{oct4-package}
\title{Salmon quantifications for condition OCT4 knockdown in mouse ESCs}
\description{
  This package provides the output of running Salmon on a set
  of 12 RNA-seq samples from King & Klose (2017).
  For more details on the data and steps used to generate the
  quantification files, please refer to the package vignette.
}
\references{
  King & Klose, "The pioneer 
  factor OCT4 requires the chromatin remodeller BRG1 to support 
  gene regulatory element function in mouse embryonic stem cells"
  eLIFE, March 2017 doi: 10.7554/eLife.22631.
}
\keyword{package}
